+++ modules/api/src/main/java/com/nisovin/shopkeepers/api/shopkeeper/ShopkeeperCreateException.java
@@ -1,8 +1,7 @@
 package com.nisovin.shopkeepers.api.shopkeeper;
 
 /**
- * This exception gets used during shopkeeper creation and loading, if the shopkeeper cannot be created due to invalid
- * or missing data.
+ * This exception is thrown when a shopkeeper cannot be created, or recreated from its save data.
  */
 public class ShopkeeperCreateException extends Exception {
 
@@ -12,9 +11,21 @@ public class ShopkeeperCreateException extends Exception {
 	 * Creates a new {@link ShopkeeperCreateException}.
 	 * 
 	 * @param message
-	 *            the exception message
+	 *            the detail message
 	 */
 	public ShopkeeperCreateException(String message) {
 		super(message);
+	}
+
+	/**
+	 * Creates a new {@link ShopkeeperCreateException}.
+	 * 
+	 * @param message
+	 *            the detail message
+	 * @param cause
+	 *            the cause
+	 */
+	public ShopkeeperCreateException(String message, Throwable cause) {
+		super(message, cause);
 	}
 }