+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopkeeper/ShopkeeperData.java
@@ -2,10 +2,10 @@ package com.nisovin.shopkeepers.shopkeeper;
 
 import com.nisovin.shopkeepers.SKShopkeepersPlugin;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopType;
-import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperCreateException;
 import com.nisovin.shopkeepers.shopobjects.ShopObjectData;
 import com.nisovin.shopkeepers.util.data.DataContainer;
 import com.nisovin.shopkeepers.util.data.DelegateDataContainer;
+import com.nisovin.shopkeepers.util.data.InvalidDataException;
 import com.nisovin.shopkeepers.util.java.StringUtils;
 
 /**
@@ -50,13 +50,13 @@ public class ShopkeeperData extends DelegateDataContainer {
 	 * Gets the shop type identifier.
 	 * 
 	 * @return the shop type identifier, not <code>null</code> or empty
-	 * @throws ShopkeeperCreateException
+	 * @throws InvalidDataException
 	 *             if the shop type identifier could not be found
 	 */
-	public final String getShopTypeId() throws ShopkeeperCreateException {
+	public final String getShopTypeId() throws InvalidDataException {
 		String shopTypeId = this.getString(DATA_KEY_SHOP_TYPE);
 		if (StringUtils.isEmpty(shopTypeId)) {
-			throw new ShopkeeperCreateException("Missing shop type id!");
+			throw new InvalidDataException("Missing shop type id!");
 		}
 		return shopTypeId;
 	}
@@ -65,14 +65,14 @@ public class ShopkeeperData extends DelegateDataContainer {
 	 * Gets the {@link ShopType}, derived from the stored {@link #getShopTypeId() shop type identifier}.
 	 * 
 	 * @return the shop type, not <code>null</code>
-	 * @throws ShopkeeperCreateException
+	 * @throws InvalidDataException
 	 *             if no valid shop type identifier could be found
 	 */
-	public final AbstractShopType<?> getShopType() throws ShopkeeperCreateException {
+	public final AbstractShopType<?> getShopType() throws InvalidDataException {
 		String shopTypeId = this.getShopTypeId();
 		AbstractShopType<?> shopType = SKShopkeepersPlugin.getInstance().getShopTypeRegistry().get(shopTypeId);
 		if (shopType == null) {
-			throw new ShopkeeperCreateException("Unknown shop type: " + shopTypeId);
+			throw new InvalidDataException("Unknown shop type: " + shopTypeId);
 		}
 		return shopType;
 	}
@@ -81,14 +81,14 @@ public class ShopkeeperData extends DelegateDataContainer {
 	 * Gets the {@link ShopObjectData}.
 	 * 
 	 * @return the shop object data, not <code>null</code>
-	 * @throws ShopkeeperCreateException
+	 * @throws InvalidDataException
 	 *             if no shop object data could be found
 	 */
-	public final ShopObjectData getShopObjectData() throws ShopkeeperCreateException {
+	public final ShopObjectData getShopObjectData() throws InvalidDataException {
 		if (shopObjectData == null) {
 			DataContainer shopObjectDataContainer = this.getContainer(DATA_KEY_SHOP_OBJECT);
 			if (shopObjectDataContainer == null) {
-				throw new ShopkeeperCreateException("Missing shop object data!");
+				throw new InvalidDataException("Missing shop object data!");
 			}
 			shopObjectData = ShopObjectData.of(shopObjectDataContainer);
 		}